Successful implementation of Mode 4 graphics

+++ cnakeGraphics.h
@@ -1,5 +1,5 @@
-void drawGameLineBetween(u16 *buffer, int x1, int x2, int y1, int y2, u16);
-void drawGameDot(u16 *buffer, int x, int y, u16 color);
+void drawGameLineBetween(u16 *buffer, int x1, int x2, int y1, int y2, u8);
+void drawGameDot(u16 *buffer, int x, int y, u8 color);
 void drawGame(u16 *buffer, Game *g);
 void drawScore(u16 *buffer, Game g);
 void drawWalls(u16 *buffer);
